+++ web-alt/src/fetching/bulkFetcher.ts
