+++ web-alt/src/index.ts
@@ -6,4 +6,4 @@ export * from "./modificationHelpers";
 export * from "./mockRestEndpoints";
 export * from "./sessionRest";
 export * from "./annotateEndpoint";
-export * from "./lsFetching";
+export * from "./fetching";
